add transform for [ESPRESSO_EVENTS] shortcode [touch: 11362]
This will automatically the shortcode when its pasted into a block (on its own) into a Events List block.  Any attributes on the pasted shortcode should get carried over as settings on the new block.

Where this is useful is if people have saved snippets they paste into content for existing EE shortcodes and try pasting it in the new Gutenberg environment.